Extract shared list/grid mode controller
Move duplicated view-mode toggle logic from CoursesFragment and ResourcesFragment into a new ListViewModeController component. The controller now handles toggle button UI state, list/grid layout manager switching, responsive grid span recalculation on layout changes, and lifecycle-safe setup/teardown, while each fragment only wires in preference getters/setters and adapter updates.
